adding new structure of the css tests + nth child tests
